Finishing quests work now

+++ src/logic/QuestLogic.java
@@ -7,25 +7,23 @@ import main.Game;
 import main.Main;
 import manager.MobManager;
 import manager.QuestManager;
+import models.Item;
 import models.Player;
 
 public class QuestLogic {
 	
 	public void updateQuestLogic(Dialogue currentDialogue) throws SlickException {
 				
-		
 		updateHalroksRatProblem(currentDialogue);
 		updateLostChrystal(currentDialogue);
-		
-
 
 	}
 	
 	private void updateHalroksRatProblem(Dialogue currentDialogue) throws SlickException {
 				
-		if(MobManager.getPlayer().isYPressed() && currentDialogue.getQuestTitle() != null && currentDialogue.getQuestTitle().equals("Halrok's Rat Problem") && !currentDialogue.isQuestStarted()) {
+		if(MobManager.getPlayer().isYPressed() && currentDialogue.getQuestTitle() != null && Game.getQuestManager().halroksRatProblem.isInactive() && currentDialogue.getQuestTitle().equals("Halrok's Rat Problem") && !currentDialogue.isQuestStarted()) {
 			
-			Game.getQuestManager().halroksRatProblem.setActive(true);
+			Game.getQuestManager().halroksRatProblem.setActive();
 						
 			Game.getQuestManager().halroksRatProblem.getNotes().add("Halrok said that he is very scared because of a rat which is currently south from here. He wants me too kill it.");
 			
@@ -35,29 +33,28 @@ public class QuestLogic {
 			MobManager.getPlayer().getCenteredText().showText(text, Main.WIDTH/2 - (text.length() * 9)/2, Main.HEIGHT/2);
 
 		}
-		System.out.println("aaaaaaaaaa");
-		System.out.println(MobManager.filthyRat.isKillEvent());
+
 		if(Game.getQuestManager().halroksRatProblem.isActive() && MobManager.filthyRat.isKillEvent()) {
-			System.out.println("KKKKKKKKKKKKKKKK");
+
 			MobManager.filthyRat.consumeKillEvent();
 			
 			Dialogue dialogue = new Dialogue();
 			dialogue.addSentence("I killed the rat.", "Hero");
-			dialogue.addSentence("Nice. Do you want 50 gold or a heath potion?", "Halrok");
+			dialogue.addSentence("Nice, good job. Here you have 50 gold.", "Halrok");
+			dialogue.addSentence("Thanks.", "Hero");
 			
-			Dialogue dialogue2 = new Dialogue();
-			dialogue2.addSentence("I want the gold.", "Hero");
-			dialogue2.addSentence("Here take it.", "Halrok");
-			
-			Dialogue dialogue3 = new Dialogue();
-			dialogue3.addSentence("I want the health potion.", "Hero");
-			dialogue3.addSentence("Here, take the health potion. But watch out, it is pretty strong stuff.", "Halrok");
-			
-			dialogue.addChildDialogue(dialogue2);
-			dialogue.addChildDialogue(dialogue3);
 			
 			MobManager.halrok.getStartingDialogues().add(dialogue);
 						
+		}
+		
+		if(MobManager.getPlayer().isYPressed() && Game.getQuestManager().halroksRatProblem.isActive() && currentDialogue.getSentences().get(0).getText().equals("I killed the rat.")) {
+			
+			MobManager.getPlayer().addExperience(Game.getQuestManager().halroksRatProblem.getExperienceReward());
+			MobManager.getPlayer().addGold(Game.getQuestManager().halroksRatProblem.getGoldReward());
+			System.out.println("KSSA");
+			Game.getQuestManager().halroksRatProblem.setFinished();
+			
 		}
 		
 	}
@@ -66,7 +63,7 @@ public class QuestLogic {
 				
 		if(currentDialogue.getQuestTitle() != null && currentDialogue.getQuestTitle().equals("Lost Chrystal") && !currentDialogue.isQuestStarted()) {
 			
-			Game.getQuestManager().lostChrystal.setActive(true);
+			Game.getQuestManager().lostChrystal.setActive();
 			
 			Game.getQuestManager().lostChrystal.getNotes().add("Ogus lost his chrystal.");
 			Game.getQuestManager().lostChrystal.getNotes().add("It should be east from here.");
